Implement logic to calculate sum
The function `run_instructions` iterates over the list of instructions,
setting the mask or updating the value in the memory. The memory is
abstracted by a `HashMap<u64,u64>` and only contains entries for
addresses that were given in the instructions.

There are two masks generated to modify the incoming value according to
the rules.