Refactor algorithm to use hashmap
Instead of scanning a long vec with possible occurrences, a hashmap is
used to store the index of the last unique number found. In case the
same number is already found, the last index is adjusted. If the number
was not previously seen `0` is stored with the current index, possibly
overriding a previous `0`.